+++ vulkanalia-sys/src/bitmasks.rs
@@ -43,47 +43,6 @@ bitflags! {
 }
 
 bitflags! {
-    /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkAccessFlagBits2KHR.html>
-    #[repr(transparent)]
-    #[derive(Default)]
-    pub struct AccessFlagBits2KHR: Flags64 {
-        const ACCESS_2_NONE = 0;
-        const ACCESS_2_INDIRECT_COMMAND_READ = 1;
-        const ACCESS_2_INDEX_READ = 1 << 1;
-        const ACCESS_2_VERTEX_ATTRIBUTE_READ = 1 << 2;
-        const ACCESS_2_UNIFORM_READ = 1 << 3;
-        const ACCESS_2_INPUT_ATTACHMENT_READ = 1 << 4;
-        const ACCESS_2_SHADER_READ = 1 << 5;
-        const ACCESS_2_SHADER_WRITE = 1 << 6;
-        const ACCESS_2_COLOR_ATTACHMENT_READ = 1 << 7;
-        const ACCESS_2_COLOR_ATTACHMENT_WRITE = 1 << 8;
-        const ACCESS_2_DEPTH_STENCIL_ATTACHMENT_READ = 1 << 9;
-        const ACCESS_2_DEPTH_STENCIL_ATTACHMENT_WRITE = 1 << 10;
-        const ACCESS_2_TRANSFER_READ = 1 << 11;
-        const ACCESS_2_TRANSFER_WRITE = 1 << 12;
-        const ACCESS_2_HOST_READ = 1 << 13;
-        const ACCESS_2_HOST_WRITE = 1 << 14;
-        const ACCESS_2_MEMORY_READ = 1 << 15;
-        const ACCESS_2_MEMORY_WRITE = 1 << 16;
-        const ACCESS_2_COMMAND_PREPROCESS_READ_NV = 1 << 17;
-        const ACCESS_2_COMMAND_PREPROCESS_WRITE_NV = 1 << 18;
-        const ACCESS_2_COLOR_ATTACHMENT_READ_NONCOHERENT_EXT = 1 << 19;
-        const ACCESS_2_CONDITIONAL_RENDERING_READ_EXT = 1 << 20;
-        const ACCESS_2_ACCELERATION_STRUCTURE_READ = 1 << 21;
-        const ACCESS_2_ACCELERATION_STRUCTURE_WRITE = 1 << 22;
-        const ACCESS_2_FRAGMENT_SHADING_RATE_ATTACHMENT_READ = 1 << 23;
-        const ACCESS_2_FRAGMENT_DENSITY_MAP_READ_EXT = 1 << 24;
-        const ACCESS_2_TRANSFORM_FEEDBACK_WRITE_EXT = 1 << 25;
-        const ACCESS_2_TRANSFORM_FEEDBACK_COUNTER_READ_EXT = 1 << 26;
-        const ACCESS_2_TRANSFORM_FEEDBACK_COUNTER_WRITE_EXT = 1 << 27;
-        const ACCESS_2_SHADER_SAMPLED_READ = 1 << 32;
-        const ACCESS_2_SHADER_STORAGE_READ = 1 << 33;
-        const ACCESS_2_SHADER_STORAGE_WRITE = 1 << 34;
-        const ACCESS_2_INVOCATION_MASK_READ_HUAWEI = 1 << 39;
-    }
-}
-
-bitflags! {
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkAccessFlags.html>
     #[repr(transparent)]
     #[derive(Default)]
@@ -123,7 +82,41 @@ bitflags! {
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkAccessFlags2KHR.html>
     #[repr(transparent)]
     #[derive(Default)]
-    pub struct AccessFlags2KHR: Flags { }
+    pub struct AccessFlags2KHR: Flags64 {
+        const NONE = 0;
+        const INDIRECT_COMMAND_READ = 1;
+        const INDEX_READ = 1 << 1;
+        const VERTEX_ATTRIBUTE_READ = 1 << 2;
+        const UNIFORM_READ = 1 << 3;
+        const INPUT_ATTACHMENT_READ = 1 << 4;
+        const SHADER_READ = 1 << 5;
+        const SHADER_WRITE = 1 << 6;
+        const COLOR_ATTACHMENT_READ = 1 << 7;
+        const COLOR_ATTACHMENT_WRITE = 1 << 8;
+        const DEPTH_STENCIL_ATTACHMENT_READ = 1 << 9;
+        const DEPTH_STENCIL_ATTACHMENT_WRITE = 1 << 10;
+        const TRANSFER_READ = 1 << 11;
+        const TRANSFER_WRITE = 1 << 12;
+        const HOST_READ = 1 << 13;
+        const HOST_WRITE = 1 << 14;
+        const MEMORY_READ = 1 << 15;
+        const MEMORY_WRITE = 1 << 16;
+        const COMMAND_PREPROCESS_READ_NV = 1 << 17;
+        const COMMAND_PREPROCESS_WRITE_NV = 1 << 18;
+        const COLOR_ATTACHMENT_READ_NONCOHERENT_EXT = 1 << 19;
+        const CONDITIONAL_RENDERING_READ_EXT = 1 << 20;
+        const ACCELERATION_STRUCTURE_READ = 1 << 21;
+        const ACCELERATION_STRUCTURE_WRITE = 1 << 22;
+        const FRAGMENT_SHADING_RATE_ATTACHMENT_READ = 1 << 23;
+        const FRAGMENT_DENSITY_MAP_READ_EXT = 1 << 24;
+        const TRANSFORM_FEEDBACK_WRITE_EXT = 1 << 25;
+        const TRANSFORM_FEEDBACK_COUNTER_READ_EXT = 1 << 26;
+        const TRANSFORM_FEEDBACK_COUNTER_WRITE_EXT = 1 << 27;
+        const SHADER_SAMPLED_READ = 1 << 32;
+        const SHADER_STORAGE_READ = 1 << 33;
+        const SHADER_STORAGE_WRITE = 1 << 34;
+        const INVOCATION_MASK_READ_HUAWEI = 1 << 39;
+    }
 }
 
 bitflags! {
@@ -1045,50 +1038,6 @@ bitflags! {
 }
 
 bitflags! {
-    /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkPipelineStageFlagBits2KHR.html>
-    #[repr(transparent)]
-    #[derive(Default)]
-    pub struct PipelineStageFlagBits2KHR: Flags64 {
-        const PIPELINE_STAGE_2_NONE = 0;
-        const PIPELINE_STAGE_2_TOP_OF_PIPE = 1;
-        const PIPELINE_STAGE_2_DRAW_INDIRECT = 1 << 1;
-        const PIPELINE_STAGE_2_VERTEX_INPUT = 1 << 2;
-        const PIPELINE_STAGE_2_VERTEX_SHADER = 1 << 3;
-        const PIPELINE_STAGE_2_TESSELLATION_CONTROL_SHADER = 1 << 4;
-        const PIPELINE_STAGE_2_TESSELLATION_EVALUATION_SHADER = 1 << 5;
-        const PIPELINE_STAGE_2_GEOMETRY_SHADER = 1 << 6;
-        const PIPELINE_STAGE_2_FRAGMENT_SHADER = 1 << 7;
-        const PIPELINE_STAGE_2_EARLY_FRAGMENT_TESTS = 1 << 8;
-        const PIPELINE_STAGE_2_LATE_FRAGMENT_TESTS = 1 << 9;
-        const PIPELINE_STAGE_2_COLOR_ATTACHMENT_OUTPUT = 1 << 10;
-        const PIPELINE_STAGE_2_COMPUTE_SHADER = 1 << 11;
-        const PIPELINE_STAGE_2_ALL_TRANSFER = 1 << 12;
-        const PIPELINE_STAGE_2_BOTTOM_OF_PIPE = 1 << 13;
-        const PIPELINE_STAGE_2_HOST = 1 << 14;
-        const PIPELINE_STAGE_2_ALL_GRAPHICS = 1 << 15;
-        const PIPELINE_STAGE_2_ALL_COMMANDS = 1 << 16;
-        const PIPELINE_STAGE_2_COMMAND_PREPROCESS_NV = 1 << 17;
-        const PIPELINE_STAGE_2_CONDITIONAL_RENDERING_EXT = 1 << 18;
-        const PIPELINE_STAGE_2_TASK_SHADER_NV = 1 << 19;
-        const PIPELINE_STAGE_2_MESH_SHADER_NV = 1 << 20;
-        const PIPELINE_STAGE_2_RAY_TRACING_SHADER = 1 << 21;
-        const PIPELINE_STAGE_2_FRAGMENT_SHADING_RATE_ATTACHMENT = 1 << 22;
-        const PIPELINE_STAGE_2_FRAGMENT_DENSITY_PROCESS_EXT = 1 << 23;
-        const PIPELINE_STAGE_2_TRANSFORM_FEEDBACK_EXT = 1 << 24;
-        const PIPELINE_STAGE_2_ACCELERATION_STRUCTURE_BUILD = 1 << 25;
-        const PIPELINE_STAGE_2_COPY = 1 << 32;
-        const PIPELINE_STAGE_2_RESOLVE = 1 << 33;
-        const PIPELINE_STAGE_2_BLIT = 1 << 34;
-        const PIPELINE_STAGE_2_CLEAR = 1 << 35;
-        const PIPELINE_STAGE_2_INDEX_INPUT = 1 << 36;
-        const PIPELINE_STAGE_2_VERTEX_ATTRIBUTE_INPUT = 1 << 37;
-        const PIPELINE_STAGE_2_PRE_RASTERIZATION_SHADERS = 1 << 38;
-        const PIPELINE_STAGE_2_SUBPASS_SHADING_HUAWEI = 1 << 39;
-        const PIPELINE_STAGE_2_INVOCATION_MASK_HUAWEI = 1 << 40;
-    }
-}
-
-bitflags! {
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkPipelineStageFlags.html>
     #[repr(transparent)]
     #[derive(Default)]
@@ -1126,7 +1075,44 @@ bitflags! {
     /// <https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkPipelineStageFlags2KHR.html>
     #[repr(transparent)]
     #[derive(Default)]
-    pub struct PipelineStageFlags2KHR: Flags { }
+    pub struct PipelineStageFlags2KHR: Flags64 {
+        const NONE = 0;
+        const TOP_OF_PIPE = 1;
+        const DRAW_INDIRECT = 1 << 1;
+        const VERTEX_INPUT = 1 << 2;
+        const VERTEX_SHADER = 1 << 3;
+        const TESSELLATION_CONTROL_SHADER = 1 << 4;
+        const TESSELLATION_EVALUATION_SHADER = 1 << 5;
+        const GEOMETRY_SHADER = 1 << 6;
+        const FRAGMENT_SHADER = 1 << 7;
+        const EARLY_FRAGMENT_TESTS = 1 << 8;
+        const LATE_FRAGMENT_TESTS = 1 << 9;
+        const COLOR_ATTACHMENT_OUTPUT = 1 << 10;
+        const COMPUTE_SHADER = 1 << 11;
+        const ALL_TRANSFER = 1 << 12;
+        const BOTTOM_OF_PIPE = 1 << 13;
+        const HOST = 1 << 14;
+        const ALL_GRAPHICS = 1 << 15;
+        const ALL_COMMANDS = 1 << 16;
+        const COMMAND_PREPROCESS_NV = 1 << 17;
+        const CONDITIONAL_RENDERING_EXT = 1 << 18;
+        const TASK_SHADER_NV = 1 << 19;
+        const MESH_SHADER_NV = 1 << 20;
+        const RAY_TRACING_SHADER = 1 << 21;
+        const FRAGMENT_SHADING_RATE_ATTACHMENT = 1 << 22;
+        const FRAGMENT_DENSITY_PROCESS_EXT = 1 << 23;
+        const TRANSFORM_FEEDBACK_EXT = 1 << 24;
+        const ACCELERATION_STRUCTURE_BUILD = 1 << 25;
+        const COPY = 1 << 32;
+        const RESOLVE = 1 << 33;
+        const BLIT = 1 << 34;
+        const CLEAR = 1 << 35;
+        const INDEX_INPUT = 1 << 36;
+        const VERTEX_ATTRIBUTE_INPUT = 1 << 37;
+        const PRE_RASTERIZATION_SHADERS = 1 << 38;
+        const SUBPASS_SHADING_HUAWEI = 1 << 39;
+        const INVOCATION_MASK_HUAWEI = 1 << 40;
+    }
 }
 
 bitflags! {